[Simulation] Move all files related to tasks into a dedicated folder
