+++ Sofa/Component/Constraint/Lagrangian/Solver/src/sofa/component/constraint/lagrangian/solver/BuiltConstraintSolver.cpp
@@ -23,8 +23,8 @@
 #include <sofa/component/constraint/lagrangian/solver/BuiltConstraintSolver.h>
 
 #include <sofa/helper/ScopedAdvancedTimer.h>
-#include <sofa/simulation/MainTaskSchedulerFactory.h>
-#include <sofa/simulation/ParallelForEach.h>
+#include <sofa/simulation/task/MainTaskSchedulerFactory.h>
+#include <sofa/simulation/task/ParallelForEach.h>
 #include <Eigen/Eigenvalues>
 
 namespace sofa::component::constraint::lagrangian::solver
