+++ Sofa/Component/LinearSystem/src/sofa/component/linearsystem/ConstantSparsityProjectionMethod.inl
@@ -24,7 +24,7 @@
 
 #include <Eigen/Sparse>
 #include <sofa/helper/ScopedAdvancedTimer.h>
-#include <sofa/simulation/MainTaskSchedulerFactory.h>
+#include <sofa/simulation/task/MainTaskSchedulerFactory.h>
 #include <sofa/simulation/ParallelSparseMatrixProduct.h>
 
 namespace sofa::component::linearsystem
