+++ Sofa/framework/Simulation/Core/CMakeLists.txt
@@ -17,8 +17,6 @@ set(HEADER_FILES
     ${SRC_ROOT}/CollisionEndEvent.h
     ${SRC_ROOT}/CollisionVisitor.h
     ${SRC_ROOT}/Colors.h
-    ${SRC_ROOT}/CpuTask.h
-    ${SRC_ROOT}/CpuTaskStatus.h
     ${SRC_ROOT}/DeactivatedNodeVisitor.h
     ${SRC_ROOT}/DefaultAnimationLoop.h
     ${SRC_ROOT}/DefaultVisualManagerLoop.h
@@ -35,7 +33,6 @@ set(HEADER_FILES
     ${SRC_ROOT}/MutationListener.h
     ${SRC_ROOT}/Node.h
     ${SRC_ROOT}/Node.inl
-    ${SRC_ROOT}/ParallelForEach.h
     ${SRC_ROOT}/ParallelSparseMatrixProduct.h
     ${SRC_ROOT}/PauseEvent.h
     ${SRC_ROOT}/PipelineImpl.h
@@ -65,19 +62,11 @@ set(HEADER_FILES
     ${SRC_ROOT}/XMLPrintVisitor.h
     ${SRC_ROOT}/init.h
     ${SRC_ROOT}/BaseSimulationExporter.h
-    ${SRC_ROOT}/TaskScheduler.h
-    ${SRC_ROOT}/TaskSchedulerFactory.h
-    ${SRC_ROOT}/TaskSchedulerRegistry.h
-    ${SRC_ROOT}/DefaultTaskScheduler.h
-    ${SRC_ROOT}/Task.h
-    ${SRC_ROOT}/InitTasks.h
     ${SRC_ROOT}/Locks.h
-    ${SRC_ROOT}/MainTaskSchedulerFactory.h
-    ${SRC_ROOT}/MainTaskSchedulerRegistry.h
     ${SRC_ROOT}/SceneCheck.h
     ${SRC_ROOT}/SceneCheckRegistry.h
     ${SRC_ROOT}/SceneCheckMainRegistry.h
-    ${SRC_ROOT}/WorkerThread.h
+
     ${SRC_ROOT}/events/BuildConstraintSystemEndEvent.h
     ${SRC_ROOT}/events/SimulationInitDoneEvent.h
     ${SRC_ROOT}/events/SimulationInitStartEvent.h
@@ -85,6 +74,7 @@ set(HEADER_FILES
     ${SRC_ROOT}/events/SimulationStartEvent.h
     ${SRC_ROOT}/events/SimulationStopEvent.h
     ${SRC_ROOT}/events/SolveConstraintSystemEndEvent.h
+
     ${SRC_ROOT}/mechanicalvisitor/MechanicalAccFromFVisitor.h
     ${SRC_ROOT}/mechanicalvisitor/MechanicalAccumulateJacobian.h
     ${SRC_ROOT}/mechanicalvisitor/MechanicalAccumulateJacobian.h
@@ -139,6 +129,19 @@ set(HEADER_FILES
     ${SRC_ROOT}/mechanicalvisitor/MechanicalVOpVisitor.h
     ${SRC_ROOT}/mechanicalvisitor/MechanicalVReallocVisitor.h
     ${SRC_ROOT}/mechanicalvisitor/MechanicalVSizeVisitor.h
+
+    ${SRC_ROOT}/task/CpuTask.h
+    ${SRC_ROOT}/task/CpuTaskStatus.h
+    ${SRC_ROOT}/task/DefaultTaskScheduler.h
+    ${SRC_ROOT}/task/InitTasks.h
+    ${SRC_ROOT}/task/MainTaskSchedulerFactory.h
+    ${SRC_ROOT}/task/MainTaskSchedulerRegistry.h
+    ${SRC_ROOT}/task/ParallelForEach.h
+    ${SRC_ROOT}/task/Task.h
+    ${SRC_ROOT}/task/TaskScheduler.h
+    ${SRC_ROOT}/task/TaskSchedulerFactory.h
+    ${SRC_ROOT}/task/TaskSchedulerRegistry.h
+    ${SRC_ROOT}/task/WorkerThread.h
 )
 
 set(SOURCE_FILES
@@ -153,8 +156,6 @@ set(SOURCE_FILES
     ${SRC_ROOT}/CollisionEndEvent.cpp
     ${SRC_ROOT}/CollisionVisitor.cpp
     ${SRC_ROOT}/Colors.cpp
-    ${SRC_ROOT}/CpuTask.cpp
-    ${SRC_ROOT}/CpuTaskStatus.cpp
     ${SRC_ROOT}/DeactivatedNodeVisitor.cpp
     ${SRC_ROOT}/DefaultAnimationLoop.cpp
     ${SRC_ROOT}/DefaultVisualManagerLoop.cpp
@@ -165,8 +166,6 @@ set(SOURCE_FILES
     ${SRC_ROOT}/InitVisitor.cpp
     ${SRC_ROOT}/IntegrateBeginEvent.cpp
     ${SRC_ROOT}/IntegrateEndEvent.cpp
-    ${SRC_ROOT}/MainTaskSchedulerRegistry.cpp
-    ${SRC_ROOT}/MainTaskSchedulerFactory.cpp
     ${SRC_ROOT}/MechanicalOperations.cpp
     ${SRC_ROOT}/MechanicalVPrintVisitor.cpp
     ${SRC_ROOT}/MechanicalVisitor.cpp
@@ -202,13 +201,7 @@ set(SOURCE_FILES
     ${SRC_ROOT}/init.cpp
     ${SRC_ROOT}/fwd.cpp
     ${SRC_ROOT}/BaseSimulationExporter.cpp
-    ${SRC_ROOT}/TaskScheduler.cpp
-    ${SRC_ROOT}/TaskSchedulerFactory.cpp
-    ${SRC_ROOT}/TaskSchedulerRegistry.cpp
-    ${SRC_ROOT}/DefaultTaskScheduler.cpp
-    ${SRC_ROOT}/Task.cpp
-    ${SRC_ROOT}/InitTasks.cpp
-    ${SRC_ROOT}/WorkerThread.cpp
+
     ${SRC_ROOT}/events/BuildConstraintSystemEndEvent.cpp
     ${SRC_ROOT}/events/SimulationInitDoneEvent.cpp
     ${SRC_ROOT}/events/SimulationInitStartEvent.cpp
@@ -271,6 +264,18 @@ set(SOURCE_FILES
     ${SRC_ROOT}/mechanicalvisitor/MechanicalVOpVisitor.cpp
     ${SRC_ROOT}/mechanicalvisitor/MechanicalVReallocVisitor.cpp
     ${SRC_ROOT}/mechanicalvisitor/MechanicalVSizeVisitor.cpp
+
+    ${SRC_ROOT}/task/CpuTask.cpp
+    ${SRC_ROOT}/task/CpuTaskStatus.cpp
+    ${SRC_ROOT}/task/DefaultTaskScheduler.cpp
+    ${SRC_ROOT}/task/InitTasks.cpp
+    ${SRC_ROOT}/task/MainTaskSchedulerFactory.cpp
+    ${SRC_ROOT}/task/MainTaskSchedulerRegistry.cpp
+    ${SRC_ROOT}/task/Task.cpp
+    ${SRC_ROOT}/task/TaskScheduler.cpp
+    ${SRC_ROOT}/task/TaskSchedulerFactory.cpp
+    ${SRC_ROOT}/task/TaskSchedulerRegistry.cpp
+    ${SRC_ROOT}/task/WorkerThread.cpp
 )
 
 sofa_find_package(Sofa.Core REQUIRED)
