+++ Sofa/framework/Simulation/Core/src/sofa/simulation/task/WorkerThread.cpp
@@ -19,8 +19,8 @@
 *                                                                             *
 * Contact information: contact@sofa-framework.org                             *
 ******************************************************************************/
-#include <sofa/simulation/WorkerThread.h>
-#include <sofa/simulation/DefaultTaskScheduler.h>
+#include <sofa/simulation/task/WorkerThread.h>
+#include <sofa/simulation/task/DefaultTaskScheduler.h>
 
 #include <cassert>
 #include <mutex>
