+++ applications/plugins/MultiThreading/src/MultiThreading/TaskSchedulerUser.cpp
@@ -20,8 +20,8 @@
 * Contact information: contact@sofa-framework.org                             *
 ******************************************************************************/
 #include <MultiThreading/TaskSchedulerUser.h>
-#include <sofa/simulation/MainTaskSchedulerFactory.h>
-#include <sofa/simulation/InitTasks.h>
+#include <sofa/simulation/task/MainTaskSchedulerFactory.h>
+#include <sofa/simulation/task/InitTasks.h>
 
 namespace multithreading
 {
