+++ applications/plugins/MultiThreading/src/MultiThreading/component/collision/detection/algorithm/ParallelBVHNarrowPhase.h
@@ -25,7 +25,7 @@
 #include <MultiThreading/TaskSchedulerUser.h>
 
 #include <sofa/component/collision/detection/algorithm/BVHNarrowPhase.h>
-#include <sofa/simulation/CpuTask.h>
+#include <sofa/simulation/task/CpuTask.h>
 #include <unordered_set>
 
 namespace multithreading::component::collision::detection::algorithm
