+++ applications/plugins/MultiThreading/src/MultiThreading/component/solidmechanics/fem/elastic/ParallelHexahedronFEMForceField.inl
@@ -22,9 +22,9 @@
 #pragma once
 
 #include <MultiThreading/component/solidmechanics/fem/elastic/ParallelHexahedronFEMForceField.h>
-#include <sofa/simulation/TaskScheduler.h>
-#include <sofa/simulation/MainTaskSchedulerFactory.h>
-#include <sofa/simulation/ParallelForEach.h>
+#include <sofa/simulation/task/TaskScheduler.h>
+#include <sofa/simulation/task/MainTaskSchedulerFactory.h>
+#include <sofa/simulation/task/ParallelForEach.h>
 
 namespace multithreading::component::forcefield::solidmechanics::fem::elastic
 {
